backend_python/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
python-dotenv
requests
pydantic
websockets
openai
google-api-python-client
beautifulsoup4
